perf(ImageServiceImpl.kt): offload thumbnail creation to a coroutine
Improve performance by executing the create and upload
thumbnail process in a separate coroutine on an IO dispatcher,
effectively offloading the work to be non-blocking. This aims
to enhance efficiency and responsiveness of the image service.
Note that further improvement is suggested by considering
using a message queue for this task in the future.